+++ kotlin-stdlib/commonMain/kotlin/com/javiersc/kotlin/stdlib/Collections.kt
@@ -126,3 +126,11 @@ internal inline fun <T> Iterable<T>.getIndexOrNull(index: Int): T? {
         }
     }
 }
+
+/** Remove all consecutive line breaks in a list of `String` */
+public fun List<String>.removeDuplicateEmptyLines(): String =
+    if (isNotEmpty()) {
+        reduce { acc: String, b: String ->
+            if (acc.lines().lastOrNull().isNullOrBlank() && b.isBlank()) acc else "$acc\n$b"
+        }
+    } else ""