+++ kotlin-stdlib/commonMain/kotlin/com/javiersc/kotlin/stdlib/Strings.kt
@@ -39,3 +39,10 @@ public inline fun String?.isNotNullNorEmpty(): Boolean {
 
 public inline val String.Companion.Empty: String
     get() = ""
+
+/** Remove all consecutive line breaks in a `String` */
+public fun String.removeDuplicateEmptyLines(): String = lines().removeDuplicateEmptyLines()
+
+/** Add an empty line as last item if it is not empty */
+public fun String.endWithNewLine(): String =
+    if (lines().lastOrNull().isNullOrBlank()) this else "$this\n"